refactor(domain): decompose in_memory_store.rs into focused submodules (KAN-767)
